# schema/openapi.json
{
  "components": {
    "schemas": {
...
      "GeneralSettingsRequest": {
        "properties": {
          "auto_remove_when_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Remove When Available"
          },
          "interval_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Seconds"
          },
          "sync_interval_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Interval Minutes"
          },
          "trending_sync_interval_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trending Sync Interval Minutes"
          }
        },
        "title": "GeneralSettingsRequest",
        "type": "object"
      },
      "GeneralSettingsResponse": {
        "properties": {
          "auto_remove_when_available": {
            "title": "Auto Remove When Available",
            "type": "boolean"
          },
          "interval_seconds": {
            "title": "Interval Seconds",
            "type": "integer"
          },
          "sync_interval_minutes": {
            "title": "Sync Interval Minutes",
            "type": "integer"
          },
          "trending_sync_interval_minutes": {
            "title": "Trending Sync Interval Minutes",
            "type": "integer"
          }
        },
        "required": [
          "interval_seconds",
          "sync_interval_minutes",
          "trending_sync_interval_minutes",
          "auto_remove_when_available"
        ],
        "title": "GeneralSettingsResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Item": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "imdb": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imdb"
          },
          "list_id": {
            "title": "List Id",
            "type": "string"
          },
          "seer_request_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seer Request Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "tmdb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb"
          },
          "trakt_id": {
            "title": "Trakt Id",
            "type": "integer"
          },
          "tvdb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tvdb"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "trakt_id",
          "type",
          "title",
          "year",
          "tmdb",
          "tvdb",
          "imdb",
          "list_id",
          "seer_request_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Item",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "All-in-One ARR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/items": {
      "get": {
        "operationId": "get_items_api_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "list",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "List"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  },
                  "title": "Response Get Items Api Items Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Items"
      }
    },
...
    "/api/items/{list_id}/{trakt_id}": {
      "delete": {
        "description": "Remove a single tracked item from its Trakt list.",
        "operationId": "delete_item_api_items__list_id___trakt_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "trakt_id",
            "required": true,
            "schema": {
              "title": "Trakt Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Item"
      }
    },
...
    "/api/posters/{media_type}/{tmdb_id}": {
      "get": {
        "description": "Serve a cached poster thumbnail, fetching it on first request.",
        "operationId": "get_poster_api_posters__media_type___tmdb_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "media_type",
            "required": true,
            "schema": {
              "title": "Media Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tmdb_id",
            "required": true,
            "schema": {
              "title": "Tmdb Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "imdb",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Imdb"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Poster"
      }
    },
...
    "/api/settings/general": {
      "get": {
        "operationId": "get_general_settings_api_settings_general_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get General Settings"
      },
      "put": {
        "operationId": "put_general_settings_api_settings_general_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put General Settings"
      }
    },
...
  }
}